Move dashboard_metadata under dashboard package
