+++ databuilder/models/dashboard/dashboard_metadata.py
